Move SdkOperationclass to client-core package
