+++ client-core/src/main/java/software/amazon/smithy/java/runtime/client/core/SdkOperation.java
@@ -3,7 +3,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package software.amazon.smithy.java.runtime.core.schema;
+package software.amazon.smithy.java.runtime.client.core;
+
+import software.amazon.smithy.java.runtime.core.schema.SdkSchema;
+import software.amazon.smithy.java.runtime.core.schema.SdkShapeBuilder;
+import software.amazon.smithy.java.runtime.core.schema.SerializableStruct;
+import software.amazon.smithy.java.runtime.core.schema.TypeRegistry;
 
 /**
  * Represents a modeled Smithy operation.
